chore: update references and naming
