--- config/tsconfig.base.json
@@ -1,32 +1,0 @@
-{
-  "extends": "./tsconfig.strict",
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2018",
-    "noImplicitAny": false,
-    "moduleResolution": "node",
-    "sourceMap": true,
-		"baseUrl": "./packages",
-    "paths": {
-			"*": ["./node_modules/*"],
-			"timeline-state-resolver": ["timeline-state-resolver"],
-			"timeline-state-resolver-types": ["timeline-state-resolver-types"]
-		},
-    "declaration": true,
-    "importHelpers": true,
-    "listFiles": false,
-    "traceResolution": false,
-    "pretty": true,
-    "lib": [
-      "es6"
-    ],
-    "types": [
-      "node",
-      "jest"
-    ]
-  },
-  "include": [
-    "../src/**/*.ts"
-  ],
-  "compileOnSave": false
-}
