--- config/tsconfig.flexible.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "strictNullChecks": true
-  }
-}
