+++ packages/input-gateway/package.json
@@ -6,28 +6,14 @@
 	"license": "MIT",
 	"contributors": [
 		{
-			"name": "Johan Nyman",
-			"email": "johan@superfly.tv",
-			"url": "http://superfly.tv"
-		},
-		{
 			"name": "Balte de Wit",
 			"email": "balte@superfly.tv",
-			"url": "http://superfly.tv"
-		},
-		{
-			"name": "Jesper Stærkær",
-			"email": "jesper@superfly.tv",
 			"url": "http://superfly.tv"
 		},
 		{
 			"name": "Jan Starzak",
 			"email": "jan@superfly.tv",
 			"url": "http://superfly.tv"
-		},
-		{
-			"name": "Stephan Nordnes Eriksen",
-			"email": "Stephanruler@gmail.com"
 		},
 		{
 			"name": "Julian Waller",
