+++ packages/input-gateway/src/config.ts
@@ -10,7 +10,6 @@ let deviceToken: string = process.env.DEVICE_TOKEN || ''
 let disableWatchdog: boolean = process.env.DISABLE_WATCHDOG === '1' || false
 let unsafeSSL: boolean = process.env.UNSAFE_SSL === '1' || false
 const certs: string[] = (process.env.CERTIFICATES || '').split(';') || []
-let disableAtemUpload: boolean = process.env.DISABLE_ATEM_UPLOAD === '1' || false // TODO: change this to be an opt-in instead
 
 let prevProcessArg = ''
 process.argv.forEach((val) => {
@@ -34,8 +33,6 @@ process.argv.forEach((val) => {
 		// arguments with no options:
 	} else if (val.match(/-disableWatchdog/i)) {
 		disableWatchdog = true
-	} else if (val.match(/-disableAtemUpload/i)) {
-		disableAtemUpload = true
 	} else if (val.match(/-unsafeSSL/i)) {
 		// Will cause the Node applocation to blindly accept all certificates. Not recommenced unless in local, controlled networks.
 		unsafeSSL = true
@@ -59,4 +56,4 @@ const config: Config = {
 	},
 }
 
-export { config, logPath, disableWatchdog, disableAtemUpload }
+export { config, logPath, disableWatchdog }
