+++ typedoc.json
@@ -1,5 +1,5 @@
 {
-	"name": "Timeline State Resolver",
+	"name": "Input Gateway",
 	"out": "docs",
 	"theme": "default",
 	"exclude": [
